\name{branch_list}
\alias{branch_list}

\title{
 get the branch vertex list
}

\description{
 get the branch vertex list
}

\usage{
  branch_list(g, nodes);
}

\arguments{
   \item{g}{graph object}
   \item{nodes}{the spine skeleton vertices of graph}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

