\name{color_bar}
\alias{color_bar}

\title{
 Plot color bar
}

\description{
 Plot color bar
}

\usage{
  color_bar(col.b.p, col.n, alpha, color.spec, v.min, v.max, 
            col.b.n=100, b.type="h", title="", col.b=NULL);
}

\arguments{
   \item{col.b.p}{color bar position}
   \item{col.n}{color parameters}
   \item{alpha}{value for transparency}
   \item{color.spec}{"rgb", "hsv" or "hcl"}
   \item{v.min}{minimum value}
   \item{v.max}{maximum value}
   \item{col.b.n}{color bar density}
   \item{b.type}{h or v: horizontal or vertical color bar}
   \item{title}{color bar title}
   \item{col.b}{background color(s)}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

