\name{ctGCluster}
\alias{ctGCluster}

\title{
 contour graph visualization
}

\description{
 contour graph visualization
}

\usage{
  ctGCluster(dat=dat, alpha=alpha, clu=clu, col1=col1, col2=col2, col3=col3, main=main);
}

\arguments{
   \item{dat}{data.frame or matrix}
   \item{alpha}{alpha value}
   \item{clu}{cluster of dat}
   \item{col1}{contour line colors}
   \item{col2}{bg colors}
   \item{col3}{point colors}
   \item{main}{main of plot}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

