\name{df2g}
\alias{df2g}

\title{
 convert data frame to igraph object with weights
}

\description{
 convert data frame to igraph object with weights
}

\usage{
  df2g(df, w);
}

\arguments{
   \item{df}{data frame, column 1,2 are gene 1 and gene 2 respectively}
   \item{w}{weights for edges of igraph}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

