\name{hcGCluster}
\alias{hcGCluster}

\title{
 Hierarchical clustering by cluster
}

\description{
 Hierarchical clustering by cluster
}

\usage{
  hcGCluster(dat, cluster);
}

\arguments{
   \item{dat}{two columns, single cell coordinates, such as tSNE output}
   \item{cluster}{cluster ID for each row of the dat}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

