\name{i2d}
\alias{i2d}

\title{
 convert 2D image into digital coordinates
}

\description{
 convert 2D image into digital coordinates
}

\usage{
  i2d(image=image, p.n=500, scale=T);
}

\arguments{
   \item{image}{image}
   \item{p.n}{number of pointer}
   \item{scale}{scale}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

