\name{list4color}
\alias{list4color}

\title{
 map colors on list
}

\description{
 map colors on list
}

\usage{
  list4color(list.in, col, n.order=NULL);
}

\arguments{
   \item{list.in}{list}
   \item{col}{colors}
   \item{n.order}{output order}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

