\name{m2i}
\alias{m2i}

\title{
 get index by cutoff from matrix of corralation coefficent or p value
}

\description{
 get index by cutoff from matrix of corralation coefficent or p value
}

\usage{
  m2i(m, type="p", cutoff=0.001);
}

\arguments{
   \item{m}{matrix, matrix of corralation coefficent or p value}
   \item{type}{p or cor: matrix of corralation coefficent or p value}
   \item{cutoff}{cutoff of corralation coefficent or p value}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

