\name{merge_cluster}
\alias{merge_cluster}

\title{
 merge cluster
}

\description{
 merge cluster
}

\usage{
  merge_cluster(dat, cutoff=5);
}

\arguments{
   \item{dat}{column 1 is x, column 2 is y and column 3 is cluster ID}
   \item{cutoff}{if cluster number < 5, the cluster will be merged into neighbor cluster}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

