\name{mst_plot}
\alias{mst_plot}

\title{
 plot a graph removed edges by weight
}

\description{
 plot a graph removed edges by weight
}

\usage{
  mst_plot(dat=dat, ...);
}

\arguments{
   \item{dat}{data.frame or matrix}
   \item{...}{more options of segments}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

