\name{plot2_hcGCluster}
\alias{plot2_hcGCluster}

\title{
 Plot hierarchical clustering by cluster in circle
}

\description{
 Plot hierarchical clustering by cluster in circle
}

\usage{
  plot2_hcGCluster(cx, cy, r1, r2, data=data, angle.start=0,  angle.end=360, 
                   tip=TRUE, tip.cex=1, edge.color = NULL, tip.color=NULL, 
                   edge.width = 1, lab.d=1, cex=1)
}

\arguments{
   \item{cx, cy}{coordinates of centre}
   \item{r1, r2}{length of radius}
   \item{data}{cluster oblect, such as hclust output}
   \item{angle.start}{start angle}
   \item{angle.end}{end angle}
   \item{tip}{plot tip}
   \item{tip.cex}{tip font size}  
   \item{edge.color}{edge colors}
   \item{tip.color}{tip.colors}
   \item{edge.width}{edge width}
   \item{lab.d}{distance from tip to cluster}
   \item{cex}{tip font size}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

