\name{preGCluster}
\alias{preGCluster}

\title{
 graph based cluster 2
}

\description{
 graph based cluster 2
}

\usage{
  preGCluster(dat=dat, check.n=4, plot.out=FALSE);
}

\arguments{
   \item{dat}{data.frame or matrix}
   \item{check.n}{output number}
   \item{plot.out}{plot.out}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

