\name{shadowtext}
\alias{shadowtext}

\title{
 plot shadow text
}

\description{
 plot shadow text
}

\usage{
  shadowtext(x, y=NULL, labels, col='white', bg='black', 
                       theta= seq(0, 2*pi, length.out=50), r=0.1, ...);
}

\arguments{
   \item{x}{x coordinate}
   \item{y}{y coordinate}
   \item{labels}{label}
   \item{col}{col}
   \item{bg}{bg}
   \item{theta}{theta}
   \item{r}{r}
   \item{...}{...}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

