\name{simGCluster}
\alias{simGCluster}

\title{
 graph simulation
}

\description{
 graph simulation
}

\usage{
  simGCluster(dat=dat, clu=clu, files=files);
}

\arguments{
   \item{dat}{data.frame or matrix}
   \item{clu}{cluster of dat}
   \item{files}{simulation files}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

