\name{simGCluster2}
\alias{simGCluster2}

\title{
 graph simulation
}

\description{
 graph simulation
}

\usage{
  simGCluster2(ref.d=ref.d, ref.col=ref.col, file.n=file.n, clu.n=1:18, clu.col=clu.col);
}

\arguments{
   \item{ref.d}{target/reference data}
   \item{ref.col}{column number of cell type/cluster in ref.d}
   \item{file.n}{simulation file names}
   \item{clu.n}{cluster number in simulation data}
   \item{clu.col}{column number of clusters in simulation data}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

