\name{smlayout}
\alias{smlayout}

\title{
 trajectory analysis
}

\description{
 trajectory analysis
}

\usage{
  smlayout(x, y, l.type="random", f=0.5, v=NULL);
}

\arguments{
   \item{x}{x vector}
   \item{y}{y vector}
   \item{l.type}{layout mathod}
   \item{f}{point fraction}
   \item{v}{index for the values that used to do the smmoth}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

