\name{smlayout2}
\alias{smlayout2}

\title{
 trajectory analysis
}

\description{
 trajectory analysis
}

\usage{
  smlayout2(level2, node, mst, coord, smlayout, l.type="all", f=1);
}

\arguments{
   \item{level2}{list for the sub-branches}
   \item{node}{trunk of tree/skeleton (mst)}
   \item{mst}{the mst from tjGCluster or tjGCluster2}
   \item{coord}{shold be the same as m in tjGCluster or tjGCluster2}
   \item{smlayout}{output of smlayout}
   \item{l.type}{layout mathod}
   \item{f}{point fractions for each branch}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

