\name{tree_list}
\alias{tree_list}

\title{
 get the tree vertex list
}

\description{
 get the tree vertex list
}

\usage{
  tree_list(g, nodes);
}

\arguments{
   \item{g}{graph object}
   \item{nodes}{the spine skeleton vertices of graph}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

