\name{treeGCluster}
\alias{treeGCluster}

\title{
 graph visualization
}

\description{
 graph visualization
}

\usage{
  treeGCluster(files=files, types=NULL, k=2:30, col.i=3, col.l=col.l, lwd.s=lwd.s, l.adj=1, ...);
}

\arguments{
   \item{files}{cluster files that should sorted by k (number of cluster)}
   \item{types}{experimental names}
   \item{k}{cluster number}
   \item{col.i}{cluster code column}
   \item{col.l}{line color}
   \item{lwd.s}{lwd range}
   \item{l.adj}{lay gap adjusting value}
   \item{...}{more options}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

