\name{tsneGCluster}
\alias{tsneGCluster}

\title{
 do tsne
}

\description{
 do tsne
}

\usage{
  tsneGCluster(dat, perplexity=30, max_iter = 1000, ...)
}

\arguments{
   \item{dat}{data.frame, row is single cell and column is for genes}
   \item{perplexity}{Rtsne package: Rtsne(perplexity)}
   \item{max_iter}{Rtsne package: Rtsne(max_iter)}
   \item{...}{Rtsne package: Rtsne(...)}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

