\name{wc_plot}
\alias{wc_plot}

\title{
 plot a graph removed edges by weight
}

\description{
 plot a graph removed edges by weight
}

\usage{
  wc_plot(dat=dat, wt=wt, vertex.size=2, vertex.label=NA, ...);
}

\arguments{
   \item{dat}{data.frame or matrix}
   \item{wt}{weight value}
   \item{vertex.size}{vertex.size of plot.igraph}
   \item{vertex.label}{vertex.label of plot.igraph}
   \item{...}{more option of plot.igraph}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

